+++ client/src/App.css
@@ -5,3 +5,6 @@
 button{
     color: #0a5e66;;
 }
+body{
+    background-color: #f0ead6;
+}
