+++ client/src/components/NotFound.js
@@ -9,7 +9,7 @@ function NotFound() {
         <h1 className="display-1">404</h1>
         <h2>Page Not Found</h2>
         <p>The page you are looking for doesn't exist.</p>
-        <Link to="/" className="btn notfound-btn">
+        <Link to="/home" className="btn notfound-btn">
           Go Back to Home
         </Link>
       </div>
