+++ client/src/styles/register.css
@@ -1,5 +1,6 @@
 .container-fluid {
     background-color: #f0ead6;
+    min-height: 100vh;
 }
 
 .btn-block {
