+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "newsapi": "^2.4.1"
+  }
+}
