+++ server/routes/userRoutes.js
@@ -1,10 +1,7 @@
 const express = require("express");
 const User = require("../models/userModel.js");
-const bcrypt = require("bcryptjs");
 const { registerUser, loginUser } = require("../controlers/authControler.js");
 const cors = require("cors");
-const generateToken = require("../utils.js");
-const newsRouter = require("../routes/newsRoutes.js");
 
 const userRouter = express.Router();
 userRouter.use(
